Switch from deck shuffling to indices
indices that move at 1, 2, 3, etc.. throughout the deck to reduce
shuffling complexity. reduces circuit size + simplifies logic and makes
it much easier to make it deterministic.

small downside is that it is not 100% even distribution in a single
battle but it's close enough and across all possible starts it's even