Improved Battle Logic UI Hooks (#14)
Adds more ui hooks for pre-ability use, post-ability-use, more granular
information about which part of an ability is used, and when energy
triggers are called.

These hooks are used in the active battle to grow/shrink the parts of an
ability that are being used at a given moment.

The UI now shows which ability is being used too, and when energy
triggers are done, with some energy flash graphic.

fixes some indexing/add_mod bugs in the compact code.