On chain battle fixes (#127)
Fixes bugs for both entering battle and advancing to next round of battle (ui flow issues)

Comments out part of the combat logic on-chain to avoid bounds check failures (hopefully fixed with compact release in future, or else we need a workaround). This is caused by things like a >= b ? a - b : x breaking when a < b making a - b negative even though that flow isn't supposed to be hit but it is still checked for some reason.

I renamed the player/enemy damage everywhere to damage_to_player etc to make it clearer what it is since enemy_damage could mean how much they're doing or how much they've received.