Dynamic witness div-mod
Resolves #114

Replaces all uses with this new witness-based implementation.

We now have additional circuits implemented on the JS side now as we had
to change many circuits to be impure due to their usage of this new
witness.

As a side-effect this fixes the mock-api's random ability always being
the same thing. I also fixed this on-chain by changing the hard-coded
0's to a `mock_rng_byte()` circuit that generates Z_256
deterministically so you probably won't notice any repetition.